move reg to pkg/slack, and fix getReactionUserIDs function
